requirements.txt
Django==5.2.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
django-jazzmin==3.0.1
Pillow==10.4.0
